requirements.txt
flask
gunicorn
flask-cors
pillow